mori: move from deepspeed/runtime/comm to deepspeed/comm
mori is a backend helper for TorchBackend (selected at backend init,
exposed via TorchBackend.all_gather_into_tensor), so it belongs under
deepspeed/comm/ alongside the other backend implementations rather
than under deepspeed/runtime/comm/ which holds higher-level comm
utilities.

This also drops the explicit deepspeed/runtime/comm/mori.py entry
from the check-torchdist exclude list, since deepspeed/comm/ is
already covered by the broader parent-directory exclude pattern.
